test(core): split runtime broker test responsibilities
